#24 Обновлен класс Notification, удалены лишние методы
